requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0
websockets>=12.0,<14.0
streamlit>=1.29.0,<2.0.0
openai>=1.6.0
pandas
numpy
plotly>=5.18.0
python-dotenv>=1.0.0
httpx>=0.25.0
